requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0  # For LSTM models
joblib>=1.3.0

# Bluetooth & Hardware
pybluez>=0.23  # Bluetooth scanning on Linux
bluepy>=1.3.0  # BLE scanning alternative
RPi.GPIO>=0.7.1  # Raspberry Pi GPIO control (only on Pi)

# Data Processing
pytz>=2023.3
python-dateutil>=2.8.2

# API & Web Server
flask>=3.0.0
flask-cors>=4.0.0
flask-restful>=0.3.10

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# Logging & Monitoring
psutil>=5.9.5  # System monitoring

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Data Validation
jsonschema>=4.19.0

# Hashing & Security
hashlib  # Built-in, but listed for clarity
secrets  # Built-in

# Time Series Analysis
statsmodels>=0.14.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
